test: extract isolated settings helper under tests support (#474)
